doan-backend/requirements.txt
cachelib==0.10.2
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
PyJWT==2.6.0
python-dotenv==0.21.1
six==1.16.0
Werkzeug==2.2.3
boto3==1.26.77
PyMySQL==1.0.2
numpy==1.24.2
pandas==2.0.0
Flask-PyMongo==2.3.0
cryptography==39.0.2
scipy==1.10.1
Faker==18.4.0
flask-testing==0.8.1
pdfkit
serverless_wsgi==3.0.2

#test
pytest==7.3.1
#mysqlclient==2.1.1
